requirements.txt
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
numpy